Adds `URLRouter` to route urls and execute actions
`URLRouter` instances are created using an `URLRouterFactory.with(entries: [URLRouterEntry])`, and accepts an array of `URLRouterEntry` instances.

An `URLRouterEntry` is formed by two parameters:
  * `URLTest`: is a method that would return `true` if a provided `URL` matches the pattern used to create the instance
  * `action` : is a closure that will be executed if the `URLTest` is positive

`URLRouterEntry` instances can be created using `URLRouterEntryFactory.with(pattern:)`.

Unit tests are provided, but `RegularExpressionTests` are left blank as a placeholder. Besides, one test inside `URLMatchingTests` is failing and is left in that state as a reminder of an specific scenario that's not covered.